+++ src/bin/cli.js
@@ -1,5 +1,4 @@
 #!/usr/bin/env node
-const fs = require("fs");
 const path = require("path");
 
 const { getArgumentsObject } = require("./arguments");
@@ -7,24 +6,16 @@ const {
   getDependencyTreeObjectFromFile
 } = require("../lib/log-treatment-service");
 const {
-  dependencyTreeToString,
-  filterDependencyTree
+  filterDependencyTree,
+  getModulesLevelInfo
 } = require("../lib/dependency-tree-service");
-const { logger } = require("./logger");
-
-function printModuleList(dependencyTreeObject) {
-  if (dependencyTreeObject && dependencyTreeObject.length) {
-    logger.log(
-      `\n#################### Module list ####################\n\n${dependencyTreeObject
-        .map(e => `\t- ${e.moduleId}`)
-        .join("\n")}\n\n`
-    );
-  } else {
-    logger.warn(
-      "The module list is empty. Either the filters are too restrictive or there's no maven dependency:tree information in log file."
-    );
-  }
-}
+const { logger } = require("../lib/logger");
+const {
+  printModuleList,
+  printSummary,
+  saveOutputToFile,
+  printModulesLevelInfo
+} = require("./cli-helper");
 
 function main() {
   const args = getArgumentsObject();
@@ -46,34 +37,26 @@ function main() {
     args.filter,
     args.exclude
   );
-  if (args.printModuleList) {
-    printModuleList(dependencyTreeObjectFiltered);
-  }
 
-  logger.info(
-    `Filtered ${dependencyTreeObjectFiltered.length} module/s vs original ${
-      dependencyTreeObject.length
-    } module/s. ${
-      args.filter ? `Filtered by "${args.filter.join(", ")}"` : ""
-    }. ${args.exclude ? `Excluded by "${args.exclude.join(", ")}"` : ""}`
+  args.printModuleList && printModuleList(dependencyTreeObjectFiltered);
+  args.artifacts &&
+    args.artifacts.length &&
+    printModulesLevelInfo(
+      getModulesLevelInfo(dependencyTreeObjectFiltered, args.artifacts)
+    );
+
+  printSummary(
+    dependencyTreeObjectFiltered,
+    dependencyTreeObject,
+    args.filter,
+    args.exclude
   );
 
-  if (!arguments.skipOutput) {
-    const outputContent = dependencyTreeToString(dependencyTreeObjectFiltered);
-    fs.writeFileSync(path.join(".", pathToOutputFile), outputContent);
-
-    logger.info(
-      `Execution output saved at ${pathToOutputFile}. Number of modules: ${
-        dependencyTreeObjectFiltered ? dependencyTreeObjectFiltered.length : 0
-      }`
-    );
-  } else {
-    logger.info(
-      `Execution output not saved due to arguments. Number of modules: ${
-        dependencyTreeObjectFiltered ? dependencyTreeObjectFiltered.length : 0
-      }`
-    );
-  }
+  saveOutputToFile(
+    dependencyTreeObjectFiltered,
+    pathToOutputFile,
+    args.skipOutput
+  );
 }
 
 if (require.main === module) {